Lift unit preference into React Context for shared state
Replaces the per-component useState in useUnitPreference with a
UnitPreferenceProvider that holds one shared state instance for the
entire recipes subtree. All components that call the hook now read from
and write to the same state, so a toggle anywhere on the page is
immediately reflected everywhere.

Also adds a window storage event listener so the preference stays in
sync across browser tabs.
